fix: move dependencies to dev and peer dependencies
Certain npm modules need to be the only version installed in an application. If a dependency is listed as a dependency of frontend-lib-learning-assistant and as a dependency of the parent application - frontend-app-learning, in this case - then there will be two copies of that module installed in the parent application. npm modules like react and @edx/frontend-platform need to be the only copy of those modules installed.

In other cases, it is not necessary to do this to enable the application to function, but it does help reduce the bundle size.

This commit moves npm modules that can be expected to be installed in the parent application from the dependencies array to the devDependencies and peerDependencies array.