Various fixes and improvements to enable installation into Learning MFE (#6)
* fix: move dependencies to dev and peer dependencies

Certain npm modules need to be the only version installed in an application. If a dependency is listed as a dependency of frontend-lib-learning-assistant and as a dependency of the parent application - frontend-app-learning, in this case - then there will be two copies of that module installed in the parent application. npm modules like react and @edx/frontend-platform need to be the only copy of those modules installed.

In other cases, it is not necessary to do this to enable the application to function, but it does help reduce the bundle size.

This commit moves npm modules that can be expected to be installed in the parent application from the dependencies array to the devDependencies and peerDependencies array.

* fix: export reducer from slice correctly

This commit fixes an error in the way we're exporting the generated slice reducer.

* fix: dispatch thunks correctly

This commit fixes a bug where thunks were not being dispatched correctly. This is fixed by using the useDispatch hook to get a referenceto the dispatch function, which is then applied to the thunk.

* fix: indexing into list when list is empty

When the messageList is empty, indexing into the array returns undefined. Reading a property of undefined causes an error, so we use optional chaining in this case.

* docs: update documentation with development installation instructions

* fix: clear currentMessage after submitting message

This commit clears the currentMessage when a learner submits their message to the chatbot.