Fix Xpert styling issues, fix lingering bugs, and make improvements. (#8)
* fix: Xpert styling

This commit includes a number of CSS fixes and changes. The fundamental issue is that the existing Xpert code used Tailwind classes, which we do not have access to in the learning MFE. Furthermore, the existing Xpert code did not make use of Bootstrap.

This commit replaces Tailwind classes with Bootstrap classes, where equivalent Bootstrap classes exist. Tailwind classes that did not have an equivalent Bootstrap class are replaced by equivalent CSS. The CSS files were converted to SCSS files.

Other changes include conditionally rendering the Sidebar component instead of relying on opacity and absolute positioning to hide and show the Sidebar, remove unncessary styling, and relying on flexbox for positioning where possible over absolute positioning.

* fix: Xpert bugs

This commit fixes a number of bugs in the Xpert. These fixes include changing the propType for the chatboxContainerRef from a string to an object, storing serializable values for dates in Redux, and correctly dispatching the addChatMessage thunk action creator.

* refactor: Xpert improvements

This commit makes a number of improvements to the Xpert code. These improvements include moving the API call logic from Xpert widget to Sidebar component to make it easier to follow how an API call is made and to avoid side effects, moving the API loading state to the Redux store to have a cleaner data flow and to make the data available to the entire component tree, and correctly scrolling to the bottom of the ChatBox when the Sidebar is opened.