fix: change CSS button element selector to class selector
This commit changes the CSS button element selector in the ToggleXpertButton to a CSS class selector (i.e. toggle). This is because the button selector was being applied in the frontend-app-learning application and causing changes to the page style.

Eventually, it may be nice to use something like CSS modules or styled components to scope our CSS to this component, but, for now, this will do.