fix: toggle button and action message remain in DOM when sidebar is open (#23)
This commit fixes a bug where the toggle button and action message remain in the DOM when the sidebar is open. Typically, they aren't visible because they're obfuscated by the open sidebar. However, on smaller screen sizes, or when the window is made smaller, the toggle button and action message are visible behind the sidebar. This commit conditionally renders the toggle button and action message only when the sidebar is closed.