feat: add additional Segment eventing for rendering toggle and closing call-to-action (#25)
In order to diagnose low usage rates, we're temporarily adding Segment events for rendering the toggle Xpert button and for closing the call-to-action message. This will help us determine whether the Xpert toggle button is being successfully rendered and whether the Xpert is being ignored by tracking whether learners close the call-to-action message at a high rate.

At a minimum, we expect to remove the Segment event for rendering the toggle Xpert button, because this will be a very noisy Segment event. We plan to leave it in for a few days just to accumulate data. We will evaluate whether to keep the call-to-action dismissal Segment event based on the data.