fix: remove Segment event for rendering toggle
In #25, we added a Segment event for rendering the Xpert toggle button. This results in test failures in the downstream frontend-app-learning. These failures are likely due to the fact that the frontend-app-learning test is unable to mock out the frontend-platform module installed in this library. This results in failures, because the analytics service is not configured in the frontend-app-learning tests. In light of that, we've decided to move the render Segment event to the Chat component in the frontend-app-learning, which wraps the Xpert component.