fix: update expiration logic to consider trials with 0 or -0 days remaining expired (#87)
This commit updates the logic that is used to determine whether an audit trial is expired. Originally, an audit trial was considered expired if the days remaining was less than 0. This change considers an audit trial expired if the days remaining is less than or equal to 0.

This is consistent with the logic used to determine whether an audit trial is expired on the backend here: https://github.com/edx/learning-assistant/blob/93d620e49447c54d6e70f7816c6823dcc5eb2534/learning_assistant/api.py#L337.

The bug this commit fixes was that learners that accessed Xpert with an audit trial that expired that day would still be shown Xpert because the days remaining was 0 or -0, and, thus, was not considered expired. Upon submitting a message, they'd receive an error, because the backend considered the audit trial expired.