fix: make Xpert LA available to paid learners and staff (#101)
This commit fixes a bug where the Xpert Learning Assistant was only appearing for learners in upgradeable modes (i.e. audit and credit). The fix was to remove conditional rendering logic that was accidentally left following a refactor to remove audit trial experimentation code. As a result, the Xpert component no longer does any conditional rendering. The conditional rendering logic is performed by the parent Chat component in the frontend-app-learning application instead.